# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Suraksha Route API - Delivery Worker Safety and Support Platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /compensation:
    get:
      operationId: listCompensationRequests
      tags: [compensation]
      summary: List all compensation requests
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected, paid]
        - name: workerId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of compensation requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CompensationRequest"
    post:
      operationId: createCompensationRequest
      tags: [compensation]
      summary: Submit a compensation request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCompensationBody"
      responses:
        "201":
          description: Compensation request created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompensationRequest"
  /compensation/{requestId}:
    get:
      operationId: getCompensationRequest
      tags: [compensation]
      summary: Get compensation request details
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Compensation request details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompensationRequest"
    put:
      operationId: updateCompensationRequest
      tags: [compensation]
      summary: Approve or reject a compensation request
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCompensationBody"
      responses:
        "200":
          description: Compensation updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompensationRequest"
  /accidents:
    get:
      operationId: listAccidents
      tags: [accidents]
      summary: List all accident reports
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [reported, investigating, resolved, closed]
        - name: workerId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of accident reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Accident"
    post:
      operationId: reportAccident
      tags: [accidents]
      summary: Report an accident
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAccidentBody"
      responses:
        "201":
          description: Accident reported
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Accident"
  /accidents/{accidentId}:
    get:
      operationId: getAccident
      tags: [accidents]
      summary: Get accident details
      parameters:
        - name: accidentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Accident details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Accident"
    put:
      operationId: updateAccident
      tags: [accidents]
      summary: Update accident status
      parameters:
        - name: accidentId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAccidentBody"
      responses:
        "200":
          description: Accident updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Accident"
...
  /complaints:
    get:
      operationId: listComplaints
      tags: [complaints]
      summary: List all complaints
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [open, investigating, resolved, dismissed]
        - name: workerId
          in: query
          schema:
            type: integer
        - name: type
          in: query
          schema:
            type: string
            enum: [unfair_deduction, forced_unsafe_work, harassment, payment_delay, other]
      responses:
        "200":
          description: List of complaints
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Complaint"
    post:
      operationId: createComplaint
      tags: [complaints]
      summary: Submit a complaint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateComplaintBody"
      responses:
        "201":
          description: Complaint created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Complaint"
  /complaints/{complaintId}:
    get:
      operationId: getComplaint
      tags: [complaints]
      summary: Get complaint details
      parameters:
        - name: complaintId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Complaint details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Complaint"
    put:
      operationId: updateComplaint
      tags: [complaints]
      summary: Update complaint status
      parameters:
        - name: complaintId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateComplaintBody"
      responses:
        "200":
          description: Complaint updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Complaint"
...
components:
  schemas:
...
    CompensationRequest:
      type: object
      properties:
        id:
          type: integer
        workerId:
          type: integer
        workerName:
          type: string
        emergencyId:
          type: integer
        emergencyType:
          type: string
        requestedAmount:
          type: number
        approvedAmount:
          type: number
        status:
          type: string
          enum: [pending, approved, rejected, paid]
        reason:
          type: string
        daysAffected:
          type: integer
        averageDailyEarning:
          type: number
        submittedAt:
          type: string
          format: date-time
        resolvedAt:
          type: string
          format: date-time
        adminNotes:
          type: string
      required:
        - id
        - workerId
        - requestedAmount
        - status
        - reason
        - daysAffected
        - submittedAt
    CreateCompensationBody:
      type: object
      properties:
        workerId:
          type: integer
        emergencyId:
          type: integer
        reason:
          type: string
        daysAffected:
          type: integer
        description:
          type: string
      required:
        - workerId
        - reason
        - daysAffected
    UpdateCompensationBody:
      type: object
      properties:
        status:
          type: string
          enum: [approved, rejected, paid]
        approvedAmount:
          type: number
        adminNotes:
          type: string
    Accident:
      type: object
      properties:
        id:
          type: integer
        workerId:
          type: integer
        workerName:
          type: string
        description:
          type: string
        location:
          type: string
        severity:
          type: string
          enum: [minor, moderate, severe, critical]
        injuryType:
          type: string
        status:
          type: string
          enum: [reported, investigating, resolved, closed]
        reportedAt:
          type: string
          format: date-time
        nextStep:
          type: string
        insuranceClaimId:
          type: integer
      required:
        - id
        - workerId
        - description
        - severity
        - status
        - reportedAt
    CreateAccidentBody:
      type: object
      properties:
        workerId:
          type: integer
        description:
          type: string
        location:
          type: string
        severity:
          type: string
          enum: [minor, moderate, severe, critical]
        injuryType:
          type: string
      required:
        - workerId
        - description
        - severity
    UpdateAccidentBody:
      type: object
      properties:
        status:
          type: string
          enum: [reported, investigating, resolved, closed]
        nextStep:
          type: string
        insuranceClaimId:
          type: integer
...
    Complaint:
      type: object
      properties:
        id:
          type: integer
        workerId:
          type: integer
        workerName:
          type: string
        type:
          type: string
          enum: [unfair_deduction, forced_unsafe_work, harassment, payment_delay, other]
        subject:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [open, investigating, resolved, dismissed]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        submittedAt:
          type: string
          format: date-time
        resolvedAt:
          type: string
          format: date-time
        adminNotes:
          type: string
        resolution:
          type: string
      required:
        - id
        - workerId
        - type
        - subject
        - description
        - status
        - priority
        - submittedAt
    CreateComplaintBody:
      type: object
      properties:
        workerId:
          type: integer
        type:
          type: string
          enum: [unfair_deduction, forced_unsafe_work, harassment, payment_delay, other]
        subject:
          type: string
        description:
          type: string
        priority:
          type: string
          enum: [low, medium, high, urgent]
      required:
        - workerId
        - type
        - subject
        - description
    UpdateComplaintBody:
      type: object
      properties:
        status:
          type: string
          enum: [investigating, resolved, dismissed]
        adminNotes:
          type: string
        resolution:
          type: string
...