# lib/api-zod/src/generated/types/createComplaintBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Suraksha Route API - Delivery Worker Safety and Support Platform
 * OpenAPI spec version: 0.1.0
 */
import type { CreateComplaintBodyPriority } from "./createComplaintBodyPriority";
import type { CreateComplaintBodyType } from "./createComplaintBodyType";

export interface CreateComplaintBody {
  workerId: number;
  type: CreateComplaintBodyType;
  subject: string;
  description: string;
  priority?: CreateComplaintBodyPriority;
}